CRS-1819: move types for simpler import
