+++ test/hmpps/components/mini-profile/miniProfile.test.ts
@@ -1,6 +1,6 @@
 import nunjucks from 'nunjucks'
 import * as cheerio from 'cheerio'
-import { MiniProfileConfig } from '../../../../src/hmpps/@types/hmpps'
+import { MiniProfileConfig } from '../../../../src/hmpps/@types'
 import { personProfileName } from '../../../../src/hmpps/utils/utils'
 import { personDateOfBirth, personStatus } from '../../../../dist/hmpps/utils/utils'
